Add the structure for Ballots

+++ api_gateway/src/main/java/org/ciphertech/api_gateway/services/vote_authority_service/repository/BallotRepository.java
@@ -3,6 +3,6 @@ package org.ciphertech.api_gateway.services.vote_authority_service.repository;
 import org.ciphertech.api_gateway.services.vote_authority_service.entity.Ballot;
 import org.springframework.data.repository.CrudRepository;
 
-public interface BallotRepository extends CrudRepository<Ballot, Integer> {
+public interface BallotRepository extends CrudRepository<Ballot, Long> {
     // You can add custom query methods if needed
 }
